requirements.txt
django==1.11
djangorestframework==3.11.1
psycopg2-binary==2.8.5
ipython
django-cors-headers==3.2.0